kyc_service/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
numpy
opencv-python
deepface
tf-keras
pydantic
langchain
langchain-openai
langchain-core
